+++ src/lib/context-providers/index.ts
@@ -1,1 +1,4 @@
-export * from './miden-context-provider'
+export * from './miden-context-provider';
+export * from './accounts-context-provider';
+export * from './files-context-provider';
+export * from './notes-context-provider';
